Raise PendingRequestsError when there are pending requests
When tests fail due to pending requests, it's often not clear which
requests are still pending. This commit introduces a new exception
type, `PendingRequestsError`, which includes a list of requests which
were pending when the pending requests timeout is triggered. This
exception inherits from `RuntimeError` so that any higher-level code
which previously rescued it as such should continue to work.

To determine which requests are pending, this commit replaces the
`Counter` class used within `Capybara::Server` with `RequestTracker`.
It serves a similar purpose, but instead of simply counting pending
requests it stores their respective Rack `env` hashes.